Admin option for auto room creation
Adds an "auto room creation" admistration option, which
makes it so if a user loads jabbr with a room url for a
room that does not exist, it is added automatically.
This allows for dynamic urls to be used that invite users
to discuss a topic without having to make the room in advance.
The setting is disabled by default.